Restructure module layout into packages
