+++ jaxdem/rl/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Part of the JaxDEM project – https://github.com/cdelv/JaxDEM
 
-from .environment import Environment
+from .environments import Environment, MultiNavigator, SingleNavigator
 from .envWrapper import vectorise_env, clip_action_env, is_wrapped, unwrap
 from .model import Model
 from .trainer import Trainer, TrajectoryData
@@ -17,4 +17,6 @@ __all__ = [
     "Trainer",
     "TrajectoryData",
     "ActionSpace",
+    "MultiNavigator",
+    "SingleNavigator",
 ]
